.gitignore追記。BGM追加。BGMクラス名を変更。ゲーム開始待機部分を追加
